review-dashboard/requirements.txt
fastapi==0.115.0
uvicorn>=0.31.1
python-frontmatter==1.1.0
pyyaml==6.0.2
